// src/types/authType.ts
export type GuardiansInfoType = {
  id: string;
  first_name: string;
  last_name: string;
  relation_with_kids: string;
  mobile: string;
  is_default: 1;
};
// ...
export type LoginForms = {
  username: string;
  password: string;
};
// ...
export type UserDataType = {
  user_id: string | number;
  first_name: string;
  last_name: string;
  email: string;
  age: string | number;
  group: string;
  levels: string;
  branch: string;
  user_profile_photo: string | null | undefined;
  guardians_info: GuardiansInfoType[];
  registration_date: string;
  ipAddress: string;
  subscription?: SubscribeResponseType;
};

export type LoginRes = {
  token: string;
  data: UserDataType;
  refresh_token: string;
};
// ...
export interface SubscribeResponseType {
  id: string;
  current_period_end: number;
  current_period_start: number;
  customer: string;
  plan: PlanType;
  status: string;
  paused: boolean;
}